fix- removed unwanted css and components
