+++ src/assests/styles/home.css
@@ -18,10 +18,6 @@
     margin: auto;
 }
 
-.quiz-type-one,
-.quiz-type-two {
-    color: var(--primary);
-}
 
 .quiz-type-one,
 .quiz-type-two,
@@ -38,11 +34,14 @@ img {
 
 img {
     position: relative;
+    opacity: 0.5;
 }
 
 .image-text {
     position: absolute;
     width: 20rem;
+    font-size: 1.5rem;
+    letter-spacing: 2px;
 }
 
 .quiz-type-one:hover,
@@ -50,10 +49,6 @@ img {
 .quiz-type-three:hover,
 img:hover {
     box-shadow: var(--all-side-shadow);
-}
-
-.quiz-type-three .image-text {
-    color: var(--theme);
 }
 
 .quiz-type .details {
