+++ src/component/component.js
@@ -1,4 +1,5 @@
 import Navbar from "./Navbar";
 import Footer from "./Footer";
+import QuizType from "./QuizType";
 
-export { Navbar, Footer };
+export { Navbar, Footer, QuizType};
